Fix critical compilation errors and iOS deployment target
CRITICAL FIXES:
- Fixed Property model constructor to make new fields optional with defaults
- Added missing PropertyType cases (studio, loft) to all switch statements
- Fixed type casting issues in mortgage calculator
- Updated iOS deployment target from 13.0 to 14.0 for Google Maps compatibility
- Fixed searchProperties method to use correct property field name
- Resolved all compilation errors preventing app build

TECHNICAL IMPROVEMENTS:
- Property model now supports address, latitude, longitude, squareFeet, propertyType
- All PropertyType enum cases properly handled in UI components
- iOS Podfile updated to require iOS 14.0 minimum
- AppFrameworkInfo.plist updated for iOS 14.0 compatibility
- Search functionality now properly filters available properties

BUILD STATUS:
- App now compiles successfully for iOS simulator
- All critical errors resolved
- Ready for real-time sync implementation

+++ lib/screens/mortgage_calculator_screen.dart
@@ -173,7 +173,7 @@ class _MortgageCalculatorScreenState extends State<MortgageCalculatorScreen>
       monthlyPayment: monthlyPayment,
       monthlyPropertyTax: monthlyPropertyTax,
       monthlyInsurance: monthlyInsurance,
-      monthlyPMI: monthlyPMI,
+      monthlyPMI: monthlyPMI.toDouble(),
       monthlyHOA: _hoaFees,
       totalMonthlyPayment: totalMonthlyPayment,
       totalInterest: totalInterest,
